Factor out the execution code to the common library.
This functionality is shared between client and server, so there is no need to
have two implementations of this code.
